requirements.txt
cython==0.25
falcon==1.1.0
gunicorn==19.6.0
psycopg2==2.6.2
sqlalchemy==1.1.4
bcrypt==1.1.1
itsdangerous==0.24
shortuuid==0.4.2
cryptography==1.7.2
cerberus==0.9
pycryptodome==3.4
